requirements.txt
pillow
opencv-python
pytesseract
python-dotenv
pandas
pyzbar
pymongo
